# Types optimizations and 234 fix