--- a/packages/core/src/columns/columnDataTypes.ts
+++ b/packages/core/src/columns/columnDataTypes.ts
@@ -7,7 +7,7 @@ export const numberMethodNames: Exclude<
 >[] = ['gt', 'gte', 'lt', 'lte', 'step', 'int', 'finite', 'safe'];
 
 // numeric column and JSON type data for validations
-export type BaseNumberData = ColumnDataBase & {
+export interface BaseNumberData extends ColumnDataBase {
   lt?: number;
   lte?: number;
   gt?: number;
@@ -16,7 +16,7 @@ export type BaseNumberData = ColumnDataBase & {
   int?: boolean;
   finite?: boolean;
   safe?: boolean;
-};
+}
 
 // method names for string columns and JSON types to generate methods' code
 export const stringMethodNames: Exclude<
@@ -81,21 +81,27 @@ export const dateMethodNames: Exclude<
 >[] = ['min', 'max'];
 
 // date column data for validations
-export type DateColumnData = ColumnDataBase & {
+export interface DateColumnData extends ColumnDataBase {
   min?: Date;
   max?: Date;
-};
+}
 
 // method names for array column and JSON type to generate methods' code
-export const arrayMethodNames: Exclude<
-  keyof ArrayMethodsData,
-  keyof ColumnDataBase
->[] = ['min', 'max', 'length', 'nonEmpty'];
+export const arrayMethodNames: (keyof ArrayMethodsData)[] = [
+  'min',
+  'max',
+  'length',
+  'nonEmpty',
+];
 
 // array column and JSON type data for validations
-export type ArrayMethodsData = ColumnDataBase & {
+export interface ArrayMethodsData {
   min?: number;
   max?: number;
   length?: number;
   nonEmpty?: boolean;
-};
+}
+
+export interface ArrayMethodsDataForBaseColumn
+  extends ColumnDataBase,
+    ArrayMethodsData {}