--- a/packages/core/src/columns/columnSchema.ts
+++ b/packages/core/src/columns/columnSchema.ts
@@ -4,51 +4,71 @@ import {
   ErrorMessages,
   NullableColumn,
 } from './columnType';
-import { EmptyObject } from '../utils';
 
-export type ColumnSchemaGetterTableClass = {
+export interface ColumnSchemaGetterTableClass {
   prototype: { columns: ColumnTypesBase };
   inputSchema(): unknown;
   querySchema(): unknown;
-};
+}
 
 export type ColumnSchemaGetterColumns<T extends ColumnSchemaGetterTableClass> =
   T['prototype']['columns'];
 
-export interface ColumnSchemaConfig {
+export interface ColumnTypeSchemaArg {
   type: unknown;
-  parse: unknown;
+  nullable<T extends ColumnTypeBase>(
+    this: T,
+  ): NullableColumn<T, unknown, unknown, unknown>;
   encode: unknown;
+  parse: unknown;
   asType: unknown;
+  errors?<T extends ColumnTypeBase>(this: T, errors: ErrorMessages): void;
+}
+
+export interface ColumnSchemaConfig extends ColumnTypeSchemaArg {
   dateAsNumber: unknown;
   dateAsDate: unknown;
-  dateMethods: EmptyObject;
   enum: unknown;
   array: unknown;
   boolean: unknown;
   buffer: unknown;
   unknown: unknown;
   never: unknown;
-  string: unknown;
+  stringSchema: unknown;
   stringMin(max: number): unknown;
   stringMax(max: number): unknown;
   stringMinMax(min: number, max: number): unknown;
-  stringMethods: EmptyObject;
   number: unknown;
   int: unknown;
-  numberMethods: EmptyObject;
   stringNumberDate: unknown;
   timeInterval: unknown;
   bit(max?: number): unknown;
   uuid: unknown;
-  nullable<T extends ColumnTypeBase>(
-    this: T,
-  ): NullableColumn<T, unknown, unknown, unknown>;
   json(): ColumnTypeBase;
   inputSchema(this: ColumnSchemaGetterTableClass): unknown;
   outputSchema(this: ColumnSchemaGetterTableClass): unknown;
   querySchema(this: ColumnSchemaGetterTableClass): unknown;
   updateSchema(this: ColumnSchemaGetterTableClass): unknown;
   pkeySchema(this: ColumnSchemaGetterTableClass): unknown;
-  errors?<T extends ColumnTypeBase>(this: T, errors: ErrorMessages): void;
+
+  smallint(): ColumnTypeBase;
+  integer(): ColumnTypeBase;
+  real(): ColumnTypeBase;
+  smallSerial(): ColumnTypeBase;
+  serial(): ColumnTypeBase;
+
+  bigint(): ColumnTypeBase;
+  decimal(precision?: number, scale?: number): ColumnTypeBase;
+  doublePrecision(): ColumnTypeBase;
+  bigSerial(): ColumnTypeBase;
+  money(): ColumnTypeBase;
+  varchar(limit?: number): ColumnTypeBase;
+  char(limit?: number): ColumnTypeBase;
+  text(min: number, max: number): ColumnTypeBase;
+  string(limit?: number): ColumnTypeBase;
+  citext(min: number, max: number): ColumnTypeBase;
+
+  date(): ColumnTypeBase;
+  timestampNoTZ(precision?: number): ColumnTypeBase;
+  timestamp(precision?: number): ColumnTypeBase;
 }