--- a/packages/core/src/columns/operators.ts
+++ b/packages/core/src/columns/operators.ts
@@ -24,4 +24,6 @@ export type OperatorToSQL<Value, Ctx> = (
 ) => string;
 
 // Base type for the object with column operators.
-export type BaseOperators = Record<string, OperatorBase<any, any>>; // eslint-disable-line @typescript-eslint/no-explicit-any
+export interface BaseOperators {
+  [K: string]: OperatorBase<any, any>; // eslint-disable-line @typescript-eslint/no-explicit-any
+}