Fixes https://github.com/romeerez/orchid-orm/issues/234 + type optimizations.

Optimizing types is quite an interesting way to spend time, TS behavior is mostly unpredictable and it's curious to find patterns in it. After collecting more shreds of evidence, I'm going to write an article about it, maybe in a month or two.

The main metric to evaluate TS performance is the "instantiations count" value that is shown when running `tsx --extendedDiagnostics`.

Before this PR:
- pqb: instantiations 6749047, memory used 513559K, check time 10.95s
- ORM: instantiations 2661113, memory used 466366K, check time 6.62s

After:
- pqb: instantiations 5180751, memory used 451966K, check time 6.81s
- ORM: instantiations 1859083, memory used 407520K, check time 5.08s

A few commits later:
- pqb: instantiations 1984344, memory used 375868K, check time 4.25s
- ORM: instantiations 890618, memory used 355881K, check time 3.93s